# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications/preferences": {
      "get": {
        "description": "Retrieve all notification preferences and settings for the authenticated user",
        "operationId": "NotificationPreferencesController_getPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferences retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get user notification preferences",
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "description": "Update one or multiple notification preference settings",
        "operationId": "NotificationPreferencesController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update notification preferences",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/preferences/mute-user": {
      "post": {
        "description": "Mute or unmute notifications from a specific user. When muted, you will not receive any notifications from this user.",
        "operationId": "NotificationPreferencesController_muteUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User mute status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid user ID or operation"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Mute or unmute notifications from a user",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/preferences/mute-keyword": {
      "post": {
        "description": "Mute or unmute notifications that contain specific keywords in their content. Useful for filtering out unwanted content.",
        "operationId": "NotificationPreferencesController_muteKeyword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MuteKeywordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Keyword mute status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid keyword or operation"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Mute or unmute notifications containing keywords",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/preferences/reset": {
      "delete": {
        "description": "Reset all notification preferences to their default values. This action cannot be undone.",
        "operationId": "NotificationPreferencesController_resetToDefaults",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferences reset to defaults successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Reset notification preferences to defaults",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/preferences/check-permission/{type}": {
      "get": {
        "description": "Check if a specific notification type would be allowed based on current preferences. Useful for debugging notification settings.",
        "operationId": "NotificationPreferencesController_checkPermission",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "description": "Notification type to check",
            "schema": {
              "enum": [
                "LIKE",
                "REPLY",
                "REPOST",
                "FOLLOW",
                "MENTION",
                "SYSTEM"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permission check completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "allowed": {
                      "type": "boolean",
                      "example": true
                    },
                    "reason": {
                      "type": "string",
                      "example": "Notification type disabled",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid notification type"
          },
          "401": {
            "description": "Unauthorized - Authentication required"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 500
                    },
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    },
                    "timestamp": {
                      "type": "string",
                      "example": "2024-01-01T00:00:00.000Z"
                    },
                    "path": {
                      "type": "string",
                      "example": "/auth/login"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Check if a notification type is allowed",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Echo Social Media API",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.echo-app.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
...
      "NotificationPreferencesResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Preferences ID"
          },
          "userId": {
            "type": "string",
            "description": "User ID"
          },
          "likes": {
            "type": "boolean",
            "description": "Receive like notifications"
          },
          "posts": {
            "type": "boolean",
            "description": "Receive post notifications"
          },
          "replies": {
            "type": "boolean",
            "description": "Receive reply notifications"
          },
          "reposts": {
            "type": "boolean",
            "description": "Receive repost notifications"
          },
          "follows": {
            "type": "boolean",
            "description": "Receive follow notifications"
          },
          "mentions": {
            "type": "boolean",
            "description": "Receive mention notifications"
          },
          "system": {
            "type": "boolean",
            "description": "Receive system notifications"
          },
          "emailDigest": {
            "type": "boolean",
            "description": "Receive email digest"
          },
          "emailSystem": {
            "type": "boolean",
            "description": "Email for system notifications"
          },
          "pushLikes": {
            "type": "boolean",
            "description": "Push notifications enabled"
          },
          "pushReplies": {
            "type": "boolean",
            "description": "Push notifications for replies"
          },
          "pushReposts": {
            "type": "boolean",
            "description": "Push notifications for reposts"
          },
          "pushFollows": {
            "type": "boolean",
            "description": "Push notifications for follows"
          },
          "pushMentions": {
            "type": "boolean",
            "description": "Push notifications for mentions"
          },
          "pushSystem": {
            "type": "boolean",
            "description": "Push notifications for system"
          },
          "soundEnabled": {
            "type": "boolean",
            "description": "Enable notification sounds"
          },
          "vibrationEnabled": {
            "type": "boolean",
            "description": "Enable vibration"
          },
          "deliveryTiming": {
            "type": "string",
            "description": "Delivery timing",
            "enum": [
              "immediate",
              "off",
              "scheduled"
            ]
          },
          "mutedUsers": {
            "description": "List of muted user IDs",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mutedKeywords": {
            "description": "List of muted keywords",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "userId",
          "likes",
          "posts",
          "replies",
          "reposts",
          "follows",
          "mentions",
          "system",
          "emailDigest",
          "emailSystem",
          "pushLikes",
          "pushReplies",
          "pushReposts",
          "pushFollows",
          "pushMentions",
          "pushSystem",
          "soundEnabled",
          "vibrationEnabled",
          "deliveryTiming",
          "mutedUsers",
          "mutedKeywords"
        ]
      },
      "UpdatePreferencesDto": {
        "type": "object",
        "properties": {
          "likes": {
            "type": "boolean",
            "description": "Enable/disable like notifications",
            "example": true
          },
          "replies": {
            "type": "boolean",
            "description": "Enable/disable reply notifications",
            "example": true
          },
          "reposts": {
            "type": "boolean",
            "description": "Enable/disable repost notifications",
            "example": true
          },
          "follows": {
            "type": "boolean",
            "description": "Enable/disable follow notifications",
            "example": true
          },
          "mentions": {
            "type": "boolean",
            "description": "Enable/disable mention notifications",
            "example": true
          },
          "system": {
            "type": "boolean",
            "description": "Enable/disable system notifications",
            "example": true
          },
          "emailLikes": {
            "type": "boolean",
            "description": "Enable/disable email notifications for likes",
            "example": false
          },
          "emailReplies": {
            "type": "boolean",
            "description": "Enable/disable email notifications for replies",
            "example": false
          },
          "emailReposts": {
            "type": "boolean",
            "description": "Enable/disable email notifications for reposts",
            "example": false
          },
          "emailFollows": {
            "type": "boolean",
            "description": "Enable/disable email notifications for follows",
            "example": false
          },
          "emailMentions": {
            "type": "boolean",
            "description": "Enable/disable email notifications for mentions",
            "example": false
          },
          "emailSystem": {
            "type": "boolean",
            "description": "Enable/disable email notifications for system messages",
            "example": true
          },
          "emailDigest": {
            "type": "boolean",
            "description": "Enable/disable email digest",
            "example": true
          },
          "pushLikes": {
            "type": "boolean",
            "description": "Enable/disable push notifications for likes",
            "example": true
          },
          "pushReplies": {
            "type": "boolean",
            "description": "Enable/disable push notifications for replies",
            "example": true
          },
          "pushReposts": {
            "type": "boolean",
            "description": "Enable/disable push notifications for reposts",
            "example": true
          },
          "pushFollows": {
            "type": "boolean",
            "description": "Enable/disable push notifications for follows",
            "example": true
          },
          "pushMentions": {
            "type": "boolean",
            "description": "Enable/disable push notifications for mentions",
            "example": true
          },
          "pushSystem": {
            "type": "boolean",
            "description": "Enable/disable push notifications for system messages",
            "example": true
          },
          "soundEnabled": {
            "type": "boolean",
            "description": "Enable/disable notification sounds",
            "example": true
          },
          "vibrationEnabled": {
            "type": "boolean",
            "description": "Enable/disable notification vibration",
            "example": true
          },
          "deliveryTiming": {
            "type": "string",
            "description": "Notification delivery timing",
            "enum": [
              "immediate",
              "digest",
              "scheduled"
            ],
            "example": "immediate"
          },
          "mutedUsers": {
            "description": "List of muted user IDs or usernames",
            "example": [
              "john_doe",
              "jane_doe"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "mutedKeywords": {
            "description": "List of muted keywords",
            "example": [
              "spam",
              "promotion"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MuteUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "description": "User ID to mute/unmute",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "mute": {
            "type": "boolean",
            "description": "True to mute, false to unmute",
            "example": true
          }
        },
        "required": [
          "userId",
          "mute"
        ]
      },
      "MuteKeywordDto": {
        "type": "object",
        "properties": {
          "keyword": {
            "type": "string",
            "description": "Keyword to mute/unmute",
            "example": "spam"
          },
          "mute": {
            "type": "boolean",
            "description": "True to mute, false to unmute",
            "example": true
          }
        },
        "required": [
          "keyword",
          "mute"
        ]
      },
...
    }
  }
}